# docs/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pasos API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/boards": {
      "get": {
        "tags": [
          "boards"
        ],
        "summary": "List Boards",
        "operationId": "list_boards_api_v1_boards_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BoardResponse"
                  },
                  "type": "array",
                  "title": "Response List Boards Api V1 Boards Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "boards"
        ],
        "summary": "Create Board Endpoint",
        "operationId": "create_board_endpoint_api_v1_boards_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/boards/{board_id}": {
      "get": {
        "tags": [
          "boards"
        ],
        "summary": "Get Board Endpoint",
        "operationId": "get_board_endpoint_api_v1_boards__board_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "board_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Board Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "boards"
        ],
        "summary": "Update Board Endpoint",
        "operationId": "update_board_endpoint_api_v1_boards__board_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "board_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Board Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "boards"
        ],
        "summary": "Delete Board Endpoint",
        "operationId": "delete_board_endpoint_api_v1_boards__board_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "board_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Board Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/boards/{board_id}/share": {
      "post": {
        "tags": [
          "boards"
        ],
        "summary": "Create Share Endpoint",
        "operationId": "create_share_endpoint_api_v1_boards__board_id__share_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "board_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Board Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AttachmentDTO": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "link",
              "image",
              "video",
              "file"
            ],
            "title": "Kind"
          },
          "url": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 1,
            "title": "Url"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "url"
        ],
        "title": "AttachmentDTO"
      },
...
      "BoardCreateRequest": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "snapshot": {
            "$ref": "#/components/schemas/BoardSnapshotDTO-Input"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "snapshot"
        ],
        "title": "BoardCreateRequest"
      },
      "BoardResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "snapshot": {
            "$ref": "#/components/schemas/BoardSnapshotDTO-Output"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "editor",
              "viewer"
            ],
            "title": "Role"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "snapshot",
          "role",
          "created_at",
          "updated_at"
        ],
        "title": "BoardResponse"
      },
      "BoardSnapshotDTO-Input": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnSnapshotDTO"
            },
            "type": "array",
            "title": "Columns"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskSnapshotDTO"
            },
            "type": "array",
            "title": "Tasks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "columns",
          "tasks"
        ],
        "title": "BoardSnapshotDTO"
      },
      "BoardSnapshotDTO-Output": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnSnapshotDTO"
            },
            "type": "array",
            "title": "Columns"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskSnapshotDTO"
            },
            "type": "array",
            "title": "Tasks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "columns",
          "tasks"
        ],
        "title": "BoardSnapshotDTO"
      },
      "BoardUpdateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "snapshot": {
            "$ref": "#/components/schemas/BoardSnapshotDTO-Input"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "snapshot"
        ],
        "title": "BoardUpdateRequest"
      },
      "ColumnSnapshotDTO": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Id"
          },
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "order": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Order"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "order"
        ],
        "title": "ColumnSnapshotDTO"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PictogramDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "url",
          "title"
        ],
        "title": "PictogramDTO"
      },
      "ShareCreateRequest": {
        "properties": {
          "permission": {
            "type": "string",
            "enum": [
              "viewer",
              "editor"
            ],
            "title": "Permission",
            "default": "viewer"
          },
          "ttl_hours": {
            "type": "integer",
            "maximum": 720.0,
            "minimum": 1.0,
            "title": "Ttl Hours",
            "default": 168
          },
          "max_uses": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Uses"
          },
          "allow_anonymous": {
            "type": "boolean",
            "title": "Allow Anonymous",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ShareCreateRequest"
      },
...
      "ShareResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "permission": {
            "type": "string",
            "enum": [
              "viewer",
              "editor"
            ],
            "title": "Permission"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "permission",
          "expires_at",
          "url"
        ],
        "title": "ShareResponse"
      },
      "TaskSnapshotDTO": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Id"
          },
          "columnId": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Columnid"
          },
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "labels": {
            "items": {
              "type": "string",
              "maxLength": 32
            },
            "type": "array",
            "title": "Labels"
          },
          "color": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "pictograms": {
            "items": {
              "$ref": "#/components/schemas/PictogramDTO"
            },
            "type": "array",
            "title": "Pictograms"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentDTO"
            },
            "type": "array",
            "title": "Attachments"
          },
          "durationSeconds": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 86400.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Durationseconds"
          },
          "createdAt": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Createdat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "columnId",
          "title",
          "createdAt"
        ],
        "title": "TaskSnapshotDTO"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}